requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.8.1
anyio==3.6.2
apache-airflow==2.4.3
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
argcomplete==2.0.0
async-timeout==4.0.3
attrs==22.1.0
Babel==2.11.0
blinker==1.5
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron_descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.1
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.4.1
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
lazy-object-proxy==1.8.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
multidict==6.1.0
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
propcache==0.2.0
psutil==5.9.4
psycopg2==2.9.10
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.1.0
text-unidecode==1.3
twilio==9.3.4
typing_extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.12
uv==0.4.25
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarl==1.15.5
